chore: add reca to ext instead of lib
