--- lib/ReservoirCellularAutomata/.gitignore
@@ -1,5 +1,0 @@
-*.jl.*.cov
-*.jl.cov
-*.jl.mem
-/docs/Manifest*.toml
-/docs/build/
