--- lib/ReservoirCellularAutomata/Project.toml
@@ -1,28 +1,0 @@
-name = "ReservoirCellularAutomata"
-uuid = "676858ad-c2be-42d3-9530-1b713c015f0f"
-authors = ["Francesco Martinuzzi"]
-version = "0.1.0"
-
-[deps]
-CellularAutomata = "878138dc-5b27-11ea-1a71-cb95d38d6b29"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-ReservoirComputing = "7c2d2b1e-3dd4-11ea-355a-8f6a8116e294"
-
-[compat]
-Aqua = "0.8"
-CellularAutomata = "0.0.2"
-Random = "1.10"
-Reexport = "1.2.2"
-ReservoirComputing = "0.11.3"
-SafeTestsets = "0.1"
-Test = "1"
-julia = "1.10"
-
-[extras]
-Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
-SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Aqua", "Test", "SafeTestsets"]
