--- lib/ReservoirCellularAutomata/README.md
@@ -1,7 +1,0 @@
-# ReservoirCellularAutomata
-
-[![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://MartinuzziFrancesco.github.io/ReservoirCellularAutomata.jl/stable/)
-[![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://MartinuzziFrancesco.github.io/ReservoirCellularAutomata.jl/dev/)
-[![Build Status](https://github.com/MartinuzziFrancesco/ReservoirCellularAutomata.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/MartinuzziFrancesco/ReservoirCellularAutomata.jl/actions/workflows/CI.yml?query=branch%3Amain)
-[![Coverage](https://codecov.io/gh/MartinuzziFrancesco/ReservoirCellularAutomata.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/MartinuzziFrancesco/ReservoirCellularAutomata.jl)
-[![Aqua](https://raw.githubusercontent.com/JuliaTesting/Aqua.jl/master/badge.svg)](https://github.com/JuliaTesting/Aqua.jl)
