--- lib/ReservoirCellularAutomata/docs/Project.toml
@@ -1,4 +1,0 @@
-[deps]
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
-ReservoirCellularAutomata = "676858ad-c2be-42d3-9530-1b713c015f0f"
