--- lib/ReservoirCellularAutomata/docs/pages.jl
@@ -1,6 +1,0 @@
-pages = [
-    "ReservoirCellularAutomata.jl" => "index.md",
-    "Solving the 5 bit memory task" => "reca_tutorials/fbmt.md",
-    "API Manual" => "api/reca.md",
-    "References" => "references.md"
-]
