--- lib/ReservoirCellularAutomata/docs/src/index.md
@@ -1,14 +1,0 @@
-```@meta
-CurrentModule = ReservoirCellularAutomata
-```
-
-# ReservoirCellularAutomata
-
-Documentation for ReservoirCellularAutomata.
-
-```@index
-```
-
-```@autodocs
-Modules = [ReservoirCellularAutomata]
-```
