--- lib/ReservoirCellularAutomata/docs/src/references.md
@@ -1,4 +1,0 @@
-# References
-
-```@bibliography
-```
