--- lib/ReservoirCellularAutomata/docs/src/refs.bib
@@ -1,28 +1,0 @@
-@misc{Yilmaz2014,
-  doi = {10.48550/ARXIV.1410.0162},
-  url = {https://arxiv.org/abs/1410.0162},
-  author = {Yilmaz,  Ozgur},
-  keywords = {Neural and Evolutionary Computing (cs.NE),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
-  title = {Reservoir Computing using Cellular Automata},
-  publisher = {arXiv},
-  year = {2014},
-  copyright = {arXiv.org perpetual,  non-exclusive license}
-}
-
-@misc{Nichele2017,
-  doi = {10.48550/ARXIV.1703.02806},
-  url = {https://arxiv.org/abs/1703.02806},
-  author = {Nichele,  Stefano and Molund,  Andreas},
-  keywords = {Neural and Evolutionary Computing (cs.NE),  Emerging Technologies (cs.ET),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
-  title = {Deep Reservoir Computing Using Cellular Automata},
-  publisher = {arXiv},
-  year = {2017},
-  copyright = {arXiv.org perpetual,  non-exclusive license}
-}
-
-@misc{Margem2017,
-  title={An experimental study on cellular automata reservoir in pathological sequence learning tasks},
-  author={Margem, Mrwan and Yilmaz, Ozg{\"u}r},
-  year={2017},
-  publisher={Jul}
-}
